Resolved mentioned changes
